--- src/bin/ChromeOS/readme.txt
@@ -1,6 +1,0 @@
-MIGHT NOT HAVE CHROME OS CAPBILITY 
-IF NOT THEN TRY TO USE LINUX IN CHROME OS
-
-THEN INSTALL THE LINUX VERSION
-
-(C) GlassWare corporation 2022 - All rights reserved.
