--- src/bin/linux/readme.txt
@@ -1,12 +1,0 @@
-YOU ARE IN LUCK, THE GLASSWARE LAUNCHER WAS BUILT ON LINUX FIRST
-AND LINUX USERS GETS UPDATES EILER THAN WINDOWS USERS
-
-JUST INSTALL THE LINUX PACKAGE AND RUN THE LAUNCHER
-YOU SOULD SEE THE SETUP, GO THRU IT
-
-THEN BOOM, YOU HAVE THE GLASSWARE LAUNCHER INSTALLED ON YOUR SYSTEM
-
-THEN YOU CAN RUN IT FROM THE COMMAND LINE WITH GWLAUNCH - in beta
-
-
-(C) GlassWare corporation 2022 - All rights reserved.
