--- src/bin/macos/readme.txt
@@ -1,10 +1,0 @@
-ILL TRY TO HAVE MACOS CAPBILITY, BUT I DONT HAVE A MAC TO TEST IT ON.
-AND PLUS I DONT KNOW HOW MANY PEOPLE ARE USING IT ON MACOS.
-
-IF YOU WANT TO USE GLASSWARE LAUNCHER ON MACOS YOU CAN TRY TO SEE IF YOU CAN GET IT TO WORK BY INSTALLING THE LINUX VERSION AND RUNNING IT ON YOUR MACOS.
-IF NOT THEN YOU CAN TRY TO INSTALL THE MACOS VERSION (IF ITS THERE) AND RUN IT ON YOUR MACOS.
-
-THANK YOU FOR UNDERSTANDING.
-
-(C) GlassWare corporation 2022 - All rights reserved.
-
