--- src/bin/windows/readme.txt
@@ -1,8 +1,0 @@
-WELCOME WINDOWS USERS, YOUR LUCKY, WINDOWS WAS THE SECOND OPERATING SYSTEM THAT WAS SORPORTED BY GLASSWARE.
-
-INSTALL THE WINDOWS 10/11 PACKAGE,
-YOU SOULD SEE THE SETUP, GO THRU THE SETUP,
-
-AND BOOM, GLASSWARE LAUNCHER IS INSTALLED ON WINDOWS 10/11.
-
-(C) GlassWare corporation 2022 - All rights reserved.
