+++ src/linux/changeHasOpen.py
