+++ src/linux/changeHasOpen2.py
