+++ src/linux/conpile.sh
