+++ src/linux/exit.cpp
