+++ src/linux/fixgwsffile.cpp
