+++ src/linux/gwsf/fixGWSFfiles.py
