+++ src/linux/gwsf/hasOpened.gwsf
