+++ src/windows/changeHasOpen.py
@@ -1,0 +1,3 @@
+fileChange = open("gwsf/hasOpened.gwsf", "w")
+fileChange.write("1")
+fileChange.close()
