+++ src/windows/changeHasOpen2.py
@@ -1,0 +1,3 @@
+fileChange = open("gwsf/hasOpened.gwsf", "w")
+fileChange.write("0")
+fileChange.close()
