+++ src/windows/checkForUpdates.py
@@ -1,0 +1,1 @@
+print("up to date")
