+++ src/windows/exit.cpp
@@ -1,0 +1,10 @@
+#include <iostream>
+#include "header.h"
+
+using namespace std;
+
+void endProgram(int exitCode) {
+    cout << "\n\nexit code: " << exitCode << endl;
+    // system("pause");
+    exit(exitCode);
+}
