+++ src/windows/gwsf/hasOpened.gwsf
@@ -1,0 +1,1 @@
+1
