+++ src/windows/gwsf/readme.txt
@@ -1,0 +1,11 @@
+GWSF = GlassWare Save Files (.GWSF)
+DO NOT EDIT THIS FILES, YOU CAN AND WILL BREAK THE PROGRAM, 
+
+
+if for any reason it breaks, do the following:
+go to GWSF recovery, it should be in the command line interface 
+
+GWSF is a simple boolean/string/int save file format, used for saveing small amounts of data.
+you will see the file format used alot in out programs.
+
+(c) 2022 glassware Corporation
