+++ src/windows/header.h
@@ -1,0 +1,5 @@
+#include <iostream>
+#include <fstream>
+
+void fixGWSF(std::string whatFile);
+void endProgram(int exitCode);
